replace Description component and stories with Markdown
